Increate size of client id range

+++ clients/status.go
@@ -2,6 +2,6 @@ package clients
 
 type Status struct {
 	Type string `json:"type"`
-	Id uint32 `json:"id"`
+	Id uint64 `json:"id"`
 }
